requirements.txt
langchain==0.2.16
langchain-core==0.2.39
langchain-openai==0.1.23
langchain-huggingface==0.0.3
qdrant-client==1.9.2
sentence-transformers==3.0.1
pydantic==2.8.2
python-dotenv==1.0.1
streamlit==1.36.0
gTTS==2.5.1
openai==1.45.0
fastapi==0.111.0
uvicorn==0.30.1